// projects/admin/src/app/models/product.model.ts
export interface Product {
    productID: number;           
    name: string;     
    description: string;        
    price: number;   
    disPrice: number;  
    quantity: number;
    categoryID: number;
    brandID: number;
    imageUrl: string;
    count: number;
    colorID: number;
    materialID: number;
  }